test.bash now checks if it run as root and start an scrub to produce an btrfs-failed-volume

+++ contrib/btrfs/test.bash
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+# check root
+if [ "$(id -u)" != "0" ]; then
+  echo "Need to be root !"
+  exit 1
+fi
+
 if [ ! -f ./contrib/btrfs/collectd_btrfs.conf ]; then
 cat <<EOF > ./contrib/btrfs/collectd_btrfs.conf
 BaseDir "/var/lib/collectd"
@@ -17,7 +23,7 @@ LoadPlugin unixsock
 
 LoadPlugin btrfs
 <Plugin btrfs>
-  Path "/tmp/btrfstest"
+  RefreshMounts "on"
 </Plugin>
 
 EOF
@@ -29,8 +35,16 @@ if [ ! -f /tmp/btrfstest.image ]; then
 
     mkdir -p /tmp/btrfstest
     mount /tmp/btrfstest.image /tmp/btrfstest
+
+    if [ "$?" != 0 ]; then
+      echo "Mount failed, do nothing ..."
+      exit 1
+    fi
+
     dd if=/dev/urandom of=/tmp/btrfstest/random.file
-    dd if=/dev/urandom bs=1M count=1 seek=50 of=/tmp/btrfstest.image
+    sync
+    dd if=/dev/urandom bs=1M count=10 seek=70 of=/tmp/btrfstest.image
+    btrfs scrub start /tmp/btrfstest
 fi
